+++ Config/scripts/transferOp.js
@@ -1,12 +1,12 @@
 
-const { createColumn, readColumn, updateColumn } = require('../database/db')
+const { readColumn } = require('../database/db')
 
 async function checkCPF(cpf, fetchID) {
 
     let id = fetchID()
 
     const database = await readColumn(`SELECT * FROM dados_clientes WHERE cpf = '${cpf}';`)
-    const loggedUser = await readColumn(`SELECT cpf FROM dados_clientes WHERE id = ${id}`)
+    const loggedUser = await readColumn(`SELECT cpf FROM dados_clientes WHERE id = ${id};`)
 
     if (database.rowCount == 0) {
         return { outcome: 400, response: "CPF não encontrado" }
@@ -24,8 +24,15 @@ async function checkCPF(cpf, fetchID) {
 
 
 
-function transferCash(cpf, value, id) {
-    console.log(cpf, value, id)
+async function transferCash(cpf, fetchID, value) {
+    
+    let id = fetchID()
+
+    const transferUser = await readColumn(`SELECT * FROM dados_clientes WHERE cpf = '${cpf}';`)
+    const loggedUser = await readColumn(`SELECT * FROM dados_clientes WHERE id = ${id};`)
+
+    return { outcome: 200, response: {transfer: transferUser.rows[0].nome, logged: loggedUser.rows[0].nome, cash: value} }
+
 }
